Make sure experience file is saved in the same folder as the engine
Some GUI's do not set the current directory of the engine to the same directory of the engine itself. This causes the experience file to be created on the same folder as the GUI itself!

This patch aims to avoid this problem alltogether by specifying a full path to the experience file at runtime.