BrainLearn 9.1
Many thanks to Omar Khalid patch on learning
Stockfish patch
Author: Stefan Geschwentner
Date: Mon Jul 6 17:56:18 2020 +0200
Timestamp: 1594050978

Add a rank based bonus for blocked pawns.

Fix for overevaluated blocked pawns on the 5th and 6th rank.
This is a rewrite of the original idea that uses only two parameters.
Thanks to rocky640 for pointing this out.