Brainlearn 20.1 Improved optimism based on Shashchess and alignement score/gui. Fixed Opening Variety
Stockfish patch
Correctly output lowerbound/upperbound in threaded searches
fixes the lowerbound/upperbound output by taking the alpha,beta bracket
into account also if a bestThread is selected that is different from the master thread.

Instead of keeping track which bounds where used in the specific search,
in this version we simply store the quality (exact, upperbound,
lowerbound) of the score along with the actual score as information on
rootMove.

closes #4239

No functional change